+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K1B933RHED69SQC1Q4AVKQ8D"
+        "01K1CNH2T0AKSJHQFM4NSX5FNY"
     )
     r"""The id of the resource"""