+++ packages/orq-rc/RELEASES.md
@@ -2159,3 +2159,13 @@ Based on:
 - [python v3.10.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.26] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.26 - packages/orq-rc
+
+## 2025-07-30 03:21:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.592.1 (2.667.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.27] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.27] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.27 - packages/orq-rc